Merge release/130.0.3 into main

+++ nekoyume/Assets/_Scripts/Game/RaidStage.cs
@@ -50,7 +50,7 @@ namespace Nekoyume.Game
         private int _turnLimit;
         private int _wave;
         private bool _isPlaying;
-        private int _currentScore;
+        private long _currentScore;
         private int _currentBossId;
 
         public SkillController SkillController { get; private set; }
@@ -71,7 +71,7 @@ namespace Nekoyume.Game
             int bossId,
             BattleLog log,
             ArenaPlayerDigest player,
-            int damageDealt,
+            long damageDealt,
             bool isNewRecord,
             bool isPractice,
             List<FungibleAssetValue> battleRewards,
@@ -106,7 +106,7 @@ namespace Nekoyume.Game
             int bossId,
             BattleLog log,
             ArenaPlayerDigest player,
-            int damageDealt,
+            long damageDealt,
             bool isNewRecord,
             bool isPractice,
             List<FungibleAssetValue> rewards,
@@ -233,7 +233,7 @@ namespace Nekoyume.Game
         }
 
         private IEnumerator CoFinish(
-            int damageDealt,
+            long damageDealt,
             bool isNewRecord,
             bool isPractice,
             List<FungibleAssetValue> rewards,
@@ -544,7 +544,7 @@ namespace Nekoyume.Game
             }
         }
 
-        public void AddScore(int score)
+        public void AddScore(long score)
         {
             _currentScore += score;
             Widget.Find<WorldBossBattle>().UpdateScore(_currentScore);
